requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.2
tzdata==2024.1
urllib3==2.2.2
zope.event==5.0
zope.interface==6.4.post2
